fix ros-industrial/staubli_experimental#32: do not skip points based on herej()
A longer description of the problem is available on the github issue,
but the basic problem is that this is not evaluated at the correct time
(i.e. when the move is about to be executed), but at the time the move
is being queued, this means having the robot on a paused state already
triggers this problem.